Update Dockerfile, change variable names for AWS access
When using the original action, no access to AWS works, because the credentials do not work.
More precisely, none are found. I suspect that the variable naming is wrong in the docker file.
The entry point script sets the AWS variables using the ones required for the action, i.e. AWS_ACCESS_KEY
and AWS_SECRET_KEY. So these should be created when the container is run. 

Actually, I don't fully understand it, because with the container environment, they should already be
usable in the entry point script. But I'm afraid they get unset by setting them in the entry point script
again from empty values. So this a try to fix it.